Fixes cmake build, buffer and queue structure
