+++ mudock/include/mudock/alpaka_implementation/alpaka_implementation.hpp
@@ -5,5 +5,6 @@
   #include <mudock/alpaka_implementation/alpaka_dot_product.hpp>
   #include <mudock/alpaka_implementation/genetic_alpaka.hpp>
   #include <mudock/alpaka_implementation/geom_transform_alpaka.hpp>
+  #include <mudock/alpaka_implementation/invoke_kernel_alpaka.hpp>
   #include <mudock/alpaka_implementation/queue_alpaka.hpp>
 #endif
